move module data-transfer to cms-client
