+++ dashbuilder/dashbuilder-client/dashbuilder-cms-client/pom.xml
@@ -88,6 +88,11 @@
 
     <dependency>
       <groupId>org.uberfire</groupId>
+      <artifactId>uberfire-js</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>org.uberfire</groupId>
       <artifactId>uberfire-server-all</artifactId>
     </dependency>
