+++ dashbuilder/dashbuilder-client/dashbuilder-cms-client/src/main/java/org/dashbuilder/client/cms/resources/i18n/ContentManagerConstants.java
@@ -1,12 +1,12 @@
 /*
  * Copyright 2016 Red Hat, Inc. and/or its affiliates.
- *  
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *  
+ *
  *    http://www.apache.org/licenses/LICENSE-2.0
- *  
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,7 +20,7 @@ import com.google.gwt.i18n.client.Messages;
 
 public interface ContentManagerConstants extends Messages {
 
-    ContentManagerConstants INSTANCE = GWT.create(ContentManagerConstants.class);
+    public static final ContentManagerConstants INSTANCE = GWT.create(ContentManagerConstants.class);
 
     String contentExplorerNew();
 
@@ -89,5 +89,23 @@ public interface ContentManagerConstants extends Messages {
     String newPerspectivePopUpViewErrorInvalidName();
 
     String newPerspectivePopUpViewErrorDuplicatedName();
+
+    String workbenchPartTitle();
+
+    String dataTransferPopUpViewTitle();
+
+    String importResultMessage(int count);
+
+    String exportOK();
+
+    String importOK();
+
+    String exportError();
+
+    String importError();
+
+    String exportText();
+
+    String importText();
 }
