+++ dashbuilder/dashbuilder-client/dashbuilder-cms-client/src/main/resources/org/dashbuilder/client/cms/resources/i18n/ContentManagerConstants.properties
@@ -1,12 +1,12 @@
 #
 # Copyright 2016 Red Hat, Inc. and/or its affiliates.
-#  
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-#  
+#
 #    http://www.apache.org/licenses/LICENSE-2.0
-#  
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -49,4 +49,14 @@ newPerspectivePopUpViewTitle=New Page ...
 newPerspectivePopUpViewErrorEmptyName=Name is mandatory
 newPerspectivePopUpViewErrorInvalidName=Invalid page name
 newPerspectivePopUpViewErrorDuplicatedName=Page name already exists
-
+workbenchPartTitle=Dashbuilder Data Transfer
+DataTransferView.import=Import
+DataTransferView.export=Export
+dataTransferPopUpViewTitle=Dashbuilder Data Transfer
+importResultMessage=Data upload successfully, {0} files imported.
+exportOK=Data exported successfully.
+importOK=Data imported successfully.
+exportError=Data export error. Check logs for more details.
+importError=Data import error. Check logs for more details.
+exportText=Export dashbuilder related data (datasets, perspectives and navigation) to a ZIP file.
+importText=Import dashbuilder related data (datasets, perspectives and navigation) from a ZIP file.
