--- dashbuilder/dashbuilder-client/dashbuilder-data-transfer-client/pom.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0"?>
-<project
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-    xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.dashbuilder</groupId>
-    <artifactId>dashbuilder-client</artifactId>
-    <version>2.22.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>dashbuilder-data-transfer-client</artifactId>
-  <packaging>jar</packaging>
-
-  <name>Dashbuilder Data Transfer Client</name>
-
-  <dependencies>
-    <dependency>
-       <groupId>com.google.gwt</groupId>
-       <artifactId>gwt-user</artifactId>
-       <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.uberfire</groupId>
-      <artifactId>uberfire-js</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.dashbuilder</groupId>
-      <artifactId>dashbuilder-common-client</artifactId>
-    </dependency>
-
-    <dependency>
-        <groupId>org.dashbuilder</groupId>
-        <artifactId>dashbuilder-services-api</artifactId>
-    </dependency>
-
-    <!-- Test Dependencies -->
-    <dependency>
-      <groupId>org.uberfire</groupId>
-      <artifactId>uberfire-testing-utils</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.google.gwt.gwtmockito</groupId>
-      <artifactId>gwtmockito</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <resources>
-      <resource>
-        <directory>src/main/java</directory>
-      </resource>
-      <resource>
-        <directory>src/main/resources</directory>
-      </resource>
-    </resources>
-  </build>
-</project>
