--- dashbuilder/dashbuilder-client/dashbuilder-data-transfer-client/src/main/java/org/dashbuilder/data/resources/i18n/DataTransferConstants.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.dashbuilder.data.resources.i18n;
-
-import com.google.gwt.core.client.GWT;
-import com.google.gwt.i18n.client.Messages;
-
-public interface DataTransferConstants extends Messages {
-
-    public static final DataTransferConstants INSTANCE = GWT.create(DataTransferConstants.class);
-
-    String workbenchPartTitle();
-    String dataTransferPopUpViewTitle();
-    String importResultMessage(int count);
-    String exportOK();
-    String importOK();
-    String exportError();
-    String importError();
-    String exportText();
-    String importText();
-}
