--- dashbuilder/dashbuilder-client/dashbuilder-data-transfer-client/src/main/resources/META-INF/ErraiApp.properties
@@ -1,16 +1,0 @@
-#
-# ErraiApp.properties
-#
-# Do not remove, even if empty!
-#
-
-# This is a marker file. When it is detected inside a JAR or at the
-# top of any classpath, the subdirectories are scanned for deployable
-# components. As such, all Errai application modules in a project
-# should contain an ErraiApp.properties at the root of all classpaths
-# that you wish to be scanned.
-#
-# There are also some configuration options that can be set in this
-# file, although it is rarely necessary. See the documentation at
-# https://docs.jboss.org/author/display/ERRAI/ErraiApp.properties
-# for details.
