--- dashbuilder/dashbuilder-client/dashbuilder-data-transfer-client/src/main/resources/META-INF/beans.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://xmlns.jcp.org/xml/ns/javaee" bean-discovery-mode="none">
-  <scan>
-    <exclude name="org.dashbuilder.data.**"/>
-  </scan>
-</beans>
