--- dashbuilder/dashbuilder-client/dashbuilder-data-transfer-client/src/main/resources/org/dashbuilder/DataTransfer.gwt.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright 2014 Red Hat, Inc. and/or its affiliates.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.4.0//EN"
-    "http://google-web-toolkit.googlecode.com/svn/tags/2.5.0/distro-source/core/src/gwt-module.dtd">
-<module>
-  <inherits name='com.google.gwt.user.User'/>
-  <inherits name="com.google.gwt.i18n.I18N"/>
-
-  <inherits name='org.jboss.errai.bus.ErraiBus'/>
-  <inherits name='org.jboss.errai.ioc.Container'/>
-
-  <inherits name="org.uberfire.UberfireClientAll"/>
-  <inherits name="org.uberfire.UberfireAPI"/>
-  <inherits name="org.uberfire.commons.UberfireCommons"/>
-
-  <inherits name="org.dashbuilder.CommonClient"/>
-  <inherits name='org.dashbuilder.ServicesAPI'/>
-
-  <inherits name="org.jboss.errai.common.ErraiCommon"/>
-  <inherits name="org.jboss.errai.bus.ErraiBus"/>
-  <inherits name="org.jboss.errai.ui.UI"/>
-
-  <source path='data'/>
-</module>
