--- dashbuilder/dashbuilder-client/dashbuilder-data-transfer-client/src/main/resources/org/dashbuilder/data/resources/i18n/DataTransferConstants.properties
@@ -1,11 +1,0 @@
-workbenchPartTitle=Dashbuilder Data Transfer
-DataTransferView.import=Import
-DataTransferView.export=Export
-dataTransferPopUpViewTitle=Dashbuilder Data Transfer
-importResultMessage=Data upload successfully, {0} files imported.
-exportOK=Data exported successfully.
-importOK=Data imported successfully.
-exportError=Data export error. Check logs for more details.
-importError=Data import error. Check logs for more details.
-exportText=Export dashbuilder related data (datasets, perspectives and navigation) to a ZIP file.
-importText=Import dashbuilder related data (datasets, perspectives and navigation) from a ZIP file.
