+++ dashbuilder/dashbuilder-shared/dashbuilder-services-api/src/main/java/org/dashbuilder/transfer/DataTransferServices.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,13 +14,19 @@
  * limitations under the License.
  */
 
-package org.dashbuilder.data;
+package org.dashbuilder.transfer;
 
-import org.jboss.errai.ioc.client.api.EntryPoint;
-import org.jboss.errai.ui.shared.api.annotations.Bundle;
+import java.util.List;
 
-@EntryPoint
-@Bundle("resources/i18n/DataTransferConstants.properties")
-public class DataTransferEntryPoint {
+import org.jboss.errai.bus.server.annotations.Remote;
 
+@Remote
+public interface DataTransferServices {
+
+    public static String FILE_PATH = "dashbuilder-data-transfer";
+    public static String EXPORT_FILE_NAME = "export.zip";
+    public static String IMPORT_FILE_NAME = "import.zip";
+
+    public String doExport() throws Exception;
+    public List<String> doImport() throws Exception;
 }
